% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulloutPanel.R
\name{createStyle}
\alias{createStyle}
\title{Helper function for creating styles}
\usage{
createStyle(...)
}
\arguments{
\item{...}{named arguments to be coerced to css styles}
}
\description{
Helper function for creating styles
}
