% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulloutPanel.R
\name{handleStyle}
\alias{handleStyle}
\title{Styles for the pullout panel handle}
\usage{
handleStyle(
  height = "100\%",
  width = "25px",
  backgroundColor = "#ddd",
  top = 0,
  ...
)
}
\arguments{
\item{height, width}{the handle height and width, in valid CSS units}

\item{backgroundColor}{the color of the handle}

\item{top}{the distance from the top of the pullout panel to position the handle}

\item{...}{further arguments to pass to the \code{style} of the handle - these are
coerced to CSS styles by converting capitals to "-" e.g. borderRadius --> border-radius}
}
\description{
Use this in the \code{handleStyle} argument of \code{pulloutPanel}
}
