% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulloutPanel.R
\name{pulloutPanel}
\alias{pulloutPanel}
\title{Create a pullout panel}
\usage{
pulloutPanel(
  ...,
  position = c("left", "right"),
  width = "500px",
  top = 0,
  handleStyle = handleStyle(),
  handleText = NULL
)
}
\arguments{
\item{...}{contents of the panel}

\item{position}{left or right}

\item{width}{the width of the expanded panel - should be a valid CSS unit}

\item{top}{the distance from the top of the parent element to place the panel -
useful for preventing overlap e.g. in dashboard pages and navbar pages}

\item{handleStyle}{styles for the panel handle - use a call to \code{handleStyle}}

\item{handleText}{optional text for the handle}
}
\description{
Use anywhere in your app ui to fix a pullout panel to either the left or the
right of that element.
}
